pyproject.toml
[project]
name = "diplo"
version = "0.1.0"
description = "AI layer between you and the outside world — messages, email, calendar"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "beeper-desktop-api>=4.3.0",
    "croniter>=6.2.2",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
